+++ test/datastore/async_methods/destroy.test.js
@@ -22,23 +22,31 @@ describe('DS.destroy(resourceName, id)', function () {
     datastore.inject('post', p1);
 
     datastore.destroy('post', 5).then(function (id) {
-      assert.equal(id, '5', 'post 5 should have been deleted');
-      assert.equal(lifecycle.beforeDestroy.callCount, 1, 'beforeDestroy should have been called');
-      assert.equal(lifecycle.afterDestroy.callCount, 1, 'afterDestroy should have been called');
-      assert.isUndefined(datastore.get('post', 5));
-      assert.equal(datastore.lastModified('post', 5), 0);
-      assert.equal(datastore.lastSaved('post', 5), 0);
-      done();
+      try {
+        assert.equal(id, '5', 'post 5 should have been deleted');
+        assert.equal(lifecycle.beforeDestroy.callCount, 1, 'beforeDestroy should have been called');
+        assert.equal(lifecycle.afterDestroy.callCount, 1, 'afterDestroy should have been called');
+        assert.isUndefined(datastore.get('post', 5));
+        assert.equal(datastore.lastModified('post', 5), 0);
+        assert.equal(datastore.lastSaved('post', 5), 0);
+        done();
+      } catch (e) {
+        done(e);
+      }
     }).catch(function (err) {
       console.error(err.stack);
       done('should not have rejected');
     });
 
     setTimeout(function () {
-      assert.equal(1, _this.requests.length);
-      assert.equal(_this.requests[0].url, 'http://test.js-data.io/posts/5');
-      assert.equal(_this.requests[0].method, 'delete');
-      _this.requests[0].respond(200, {'Content-Type': 'text/plain'}, '5');
+      try {
+        assert.equal(1, _this.requests.length);
+        assert.equal(_this.requests[0].url, 'http://test.js-data.io/posts/5');
+        assert.equal(_this.requests[0].method, 'delete');
+        _this.requests[0].respond(200, {'Content-Type': 'text/plain'}, '5');
+      } catch (e) {
+        done(e);
+      }
     }, 30);
   });
   it('should handle nested resources', function (done) {
@@ -102,6 +110,42 @@ describe('DS.destroy(resourceName, id)', function () {
       assert.equal(_this.requests[0].url, 'http://test.js-data.io/user/4/comment/5');
       assert.equal(_this.requests[0].method, 'delete');
       _this.requests[0].respond(204);
+    }, 30);
+  });
+  it('should eager eject', function (done) {
+    var _this = this;
+
+    datastore.inject('post', p1);
+
+    datastore.destroy('post', 5, { eagerEject: true }).then(function (id) {
+      try {
+        assert.equal(id, '5', 'post 5 should have been deleted');
+        assert.equal(lifecycle.beforeDestroy.callCount, 1, 'beforeDestroy should have been called');
+        assert.equal(lifecycle.afterDestroy.callCount, 1, 'afterDestroy should have been called');
+        assert.isUndefined(datastore.get('post', 5));
+        assert.equal(datastore.lastModified('post', 5), 0);
+        assert.equal(datastore.lastSaved('post', 5), 0);
+        done();
+      } catch (e) {
+        done(e);
+      }
+    }).catch(function (err) {
+      console.error(err.stack);
+      done('should not have rejected');
+    });
+
+    setTimeout(function () {
+      assert.isUndefined(datastore.get('post', 5));
+      setTimeout(function () {
+        try {
+          assert.equal(1, _this.requests.length);
+          assert.equal(_this.requests[0].url, 'http://test.js-data.io/posts/5');
+          assert.equal(_this.requests[0].method, 'delete');
+          _this.requests[0].respond(200, {'Content-Type': 'text/plain'}, '5');
+        } catch (e) {
+          done(e);
+        }
+      }, 30);
     }, 30);
   });
 });